// frontend/src/shared/design-tokens/generated/tokens.ts
export const EDITOR_THEME_COLOR_TOKENS = {
  light: {
    pageBg: '#f4f6f8',
    workspaceBg: '#ffffff',
    panelBg: '#f8fafc',
    panelMutedBg: '#eef2f6',
    diagramBg: '#ffffff',
    diagramGridDot: '#d9e2ec',
    border: '#d9e2ec',
    borderControl: '#b9c6d3',
    borderStrong: '#7b8794',
    text: '#2f3742',
    textStrong: '#1f2933',
    textMuted: '#4f5b67',
    textSubtle: '#6b7785',
    textFaint: '#9aa6b2',
    title: '#1f2933',
    controlHoverBg: '#eef2f6',
    controlHoverSoftBg: '#f8fafc',
    controlActiveBg: '#3f8f8f',
    controlActiveHoverBg: '#327a7a',
    controlActiveText: '#ffffff',
    brand: '#3f8f8f',
    tableHeaderBg: '#eef2f6',
    tableHeaderActiveBg: '#d9e2ec',
    tableHeaderMuted: '#6b7785',
    tableBorder: '#b9c6d3',
    tableShadow: 'rgb(31 41 51 / 10%)',
    tableActiveShadow: 'rgb(31 41 51 / 15%)',
    floatingShadow: 'rgb(31 41 51 / 12%)',
    tableActiveAccent: '#b7791f',
    tableActiveRing: 'rgb(183 121 31 / 24%)',
    relationLine: '#3f8f8f',
    relationSource: '#b7791f',
    relationReference: '#3f8f8f',
    referenceAccent: '#3f8f8f',
    referenceBg: '#e6f4f4',
    sourceAccent: '#b7791f',
    sourceBg: '#fff7e8',
    columnBorder: '#e6ebf0',
    badgeBg: '#e6f4f4',
    badgeText: '#3f8f8f',
    handleBg: '#3f8f8f',
    handleBorder: '#b9c6d3',
    diagnosticsEmptyBorder: '#d9e2ec',
    diagnosticsEmptyBg: '#f8fafc',
    diagnosticsEmptyText: '#4f5b67',
    diagnosticsErrorBorder: '#e53e3e',
    diagnosticsErrorBg: '#fff5f5',
    diagnosticsText: '#1f2933',
    diagnosticsMuted: '#4f5b67',
    error: '#c53030',
    warning: '#b7791f',
  },
  'light-solarized': {
    pageBg: '#f6f1e7',
    workspaceBg: '#fffaf0',
    panelBg: '#fbf6ec',
    panelMutedBg: '#eee5d6',
    diagramBg: '#f8f3ea',
    diagramGridDot: '#dfd3c0',
    border: '#dfd3c0',
    borderControl: '#c9bba5',
    borderStrong: '#928374',
    text: '#3c3836',
    textStrong: '#282828',
    textMuted: '#6f6255',
    textSubtle: '#8a7a68',
    textFaint: '#a99b88',
    title: '#3c3836',
    controlHoverBg: '#dfd3c0',
    controlHoverSoftBg: '#eee5d6',
    controlActiveBg: '#4f8a8b',
    controlActiveHoverBg: '#3f7778',
    controlActiveText: '#ffffff',
    brand: '#4f8a8b',
    tableHeaderBg: '#eee5d6',
    tableHeaderActiveBg: '#dfd3c0',
    tableHeaderMuted: '#8a7a68',
    tableBorder: '#c9bba5',
    tableShadow: 'rgb(60 56 54 / 13%)',
    tableActiveShadow: 'rgb(60 56 54 / 18%)',
    floatingShadow: 'rgb(60 56 54 / 14%)',
    tableActiveAccent: '#b57614',
    tableActiveRing: 'rgb(181 118 20 / 24%)',
    relationLine: '#4f8a8b',
    relationSource: '#b57614',
    relationReference: '#4f8a8b',
    referenceAccent: '#4f8a8b',
    referenceBg: '#e3f0ef',
    sourceAccent: '#b57614',
    sourceBg: '#f7ead2',
    columnBorder: '#eadfce',
    badgeBg: '#e3f0ef',
    badgeText: '#4f8a8b',
    handleBg: '#4f8a8b',
    handleBorder: '#c9bba5',
    diagnosticsEmptyBorder: '#dfd3c0',
    diagnosticsEmptyBg: '#fbf6ec',
    diagnosticsEmptyText: '#6f6255',
    diagnosticsErrorBorder: '#c7605a',
    diagnosticsErrorBg: '#fae7e4',
    diagnosticsText: '#282828',
    diagnosticsMuted: '#6f6255',
    error: '#9d0006',
    warning: '#b57614',
  },
  dark: {
    pageBg: '#1b1b1b',
    workspaceBg: '#282828',
    panelBg: '#282828',
    panelMutedBg: '#32302f',
    diagramBg: '#1b1b1b',
    diagramGridDot: '#45403d',
    border: '#45403d',
    borderControl: '#504945',
    borderStrong: '#928374',
    text: '#d4be98',
    textStrong: '#ddc7a1',
    textMuted: '#a89984',
    textSubtle: '#928374',
    textFaint: '#7c6f64',
    title: '#d8a657',
    controlHoverBg: '#45403d',
    controlHoverSoftBg: '#32302f',
    controlActiveBg: '#7daea3',
    controlActiveHoverBg: '#89b482',
    controlActiveText: '#ffffff',
    brand: '#7daea3',
    tableHeaderBg: '#32302f',
    tableHeaderActiveBg: '#45403d',
    tableHeaderMuted: '#a89984',
    tableBorder: '#504945',
    tableShadow: 'rgb(0 0 0 / 26%)',
    tableActiveShadow: 'rgb(0 0 0 / 32%)',
    floatingShadow: 'rgb(0 0 0 / 28%)',
    tableActiveAccent: '#d8a657',
    tableActiveRing: 'rgb(216 166 87 / 24%)',
    relationLine: '#d4be98',
    relationSource: '#d8a657',
    relationReference: '#7daea3',
    referenceAccent: '#7daea3',
    referenceBg: '#374141',
    sourceAccent: '#d8a657',
    sourceBg: '#4f422e',
    columnBorder: '#32302f',
    badgeBg: '#374141',
    badgeText: '#7daea3',
    handleBg: '#7daea3',
    handleBorder: '#504945',
    diagnosticsEmptyBorder: '#45403d',
    diagnosticsEmptyBg: '#32302f',
    diagnosticsEmptyText: '#a89984',
    diagnosticsErrorBorder: '#633936',
    diagnosticsErrorBg: '#402120',
    diagnosticsText: '#ddc7a1',
    diagnosticsMuted: '#a89984',
    error: '#ea6962',
    warning: '#d8a657',
  },
} as const